+++ src/lib/models.ts
@@ -256,6 +256,7 @@ export type Prefs = {
 	steamLibraryDir: string | null;
 	dataDir: string;
 	cacheDir: string;
+	sendTelementary: boolean;
 	fetchModsAutomatically: boolean;
 	zoomFactor: number;
 	gamePrefs: Map<string, GamePrefs>;